Have to split tests into two binaries
