+++ tests/test_misc.rs
@@ -3,14 +3,6 @@ use slog::{o, Drain, Logger};
 use slog_stdlog::StdLog;
 
 #[test]
-fn test_default_logger() {
-    let logger = GlobalLogger::get();
-    let pairs = logger.list();
-    let expected = String::from("(zetta_version)");
-    let actual = format!("{:?}", pairs);
-    assert_eq!(expected, actual);
-}
-#[test]
 fn test_not_default_logger() {
     let root = Logger::root(StdLog.fuse(), o!("wat" => "wat"));
     GlobalLogger::setup(&root).unwrap();
